move database modules to separate subpackage
