+++ inperso/database/__init__.py
@@ -1,0 +1,6 @@
+from . import read, write
+
+__all__ = [
+    "read",
+    "write",
+]
